==> gpa-poc-churn/notebooks/00-ia-alinhamentodemo.ipynb <==
import os
import sys
import numpy as np
import pandas as pd
import seaborn as sn
import matplotlib.pyplot as plt
from google.cloud import storage

prefix = 'gs://gpa-churn/'

## Preprocessing data from STAGING

# listing objects in blob storage bucket
storage_client = storage.Client()
obj_list = storage_client.list_blobs('gpa-churn')
obj_list = [i.name for i in obj_list if 'STAGING/' in i.name]

dict_tables = {
    'registro':[i for i in obj_list if 'cadastro' in i],
    'stix':[i for i in obj_list if 'stix_optin' in i],
    'ativacao':[i for i in obj_list if 'ativacao_resgate_stix' in i],
    'email':[i for i in obj_list if 'optin_email' in i],
    'itens':[i for i in obj_list if 'relevanc_item' in i],
    'loja':[i for i in obj_list if 'relevanc_store' in i],
    'produto':[i for i in obj_list if 'relevanc_product' in i],
    'desconto':[i for i in obj_list if 'ativ_meu_desconto' in i],
    'vendas':[i for i in obj_list if 'mov_vend_crm' in i]
}

dict_tables

---

### Registro

Se houver clientes com mais de uma entrada de endereço, o que fazer? 
É possível mapear o momento em que aconteceu a mudança de endereço no cadastro? Manter primeira/última entrada?

* A atualização sobrescreve os valores anteriores (keep='last')
* data em que o cliente se cadastrou no grupo extra/...

df_list = []
for file in dict_tables['registro']:
    try:
        local_df = pd.read_csv(prefix + file, sep=';')
    except:
        local_df = pd.read_csv(prefix + file, sep='\n')
        column_list = local_df.columns[0].split(';')
        local_df = local_df[local_df.columns[0]].str.split(';', expand=True)
        local_df.columns = column_list
    df_list.append(local_df)
    print('added file: ', file)
print('-'*10)

registro = pd.concat(df_list, axis=0)
registro = registro.dropna(subset=['datcadorigem'])
registro.sort_values(by=['datcadorigem'], inplace=True)
registro

registro['datcadorigem'].value_counts()

mask_2021 = registro['datcadorigem']>='2021-01-01'
mask_2022 = registro['datcadorigem']<='2022-05-05'
registro[mask_2021 & mask_2022]

# mask_2022 = registro['datcadorigem']>='2022-05-05'
# registro[mask_2021 & mask_2022]

---

### Stix

O campo dt_cadastro_origem (stix) é, de fato, a mesma informação de datcadorigem do dataframe 'registro'?

* data em que o cliente fez cadastro no programa stix
    * oct 2020

df_list = []
for file in dict_tables['stix']:
    try:
        local_df = pd.read_csv(prefix + file, sep=';')
    except:
        local_df = pd.read_csv(prefix + file, sep='\n')
        column_list = local_df.columns[0].split(';')
        local_df = local_df[local_df.columns[0]].str.split(';', expand=True)
        local_df.columns = column_list
    df_list.append(local_df)
    print('added file: ', file)
print('-'*10)

stix = pd.concat(df_list, axis=0)
stix = stix.dropna(subset=['dt_cadastro_origem'])
stix.sort_values(by=['dt_cadastro_origem'], inplace=True)
stix

stix['dt_cadastro_origem'].value_counts()

mask_2021 = stix['dt_cadastro_origem']>='2021-01-01'
mask_2022 = stix['dt_cadastro_origem']<='2022-05-05'
stix[mask_2021 & mask_2022]

---

### Vendas - utilização do desconto

chave = [id_cliente, num_cupom, 'cod_plu_orig']

Sem dados históricos.

* Tabela de ativação do desconto
* FLG_VEND_MEU_DESCT - uso do desconto
* VAL_DESCT_MEU_DESCT - valor do desconto

df_list = []
for file in dict_tables['vendas']:
    try:
        local_df = pd.read_csv(prefix + file, sep=';')
    except:
        local_df = pd.read_csv(prefix + file, sep='\n')
        column_list = local_df.columns[0].split(';')
        local_df = local_df[local_df.columns[0]].str.split(';', expand=True)
        local_df.columns = column_list
    df_list.append(local_df)
    print('added file: ', file)
print('-'*10)

vendas = pd.concat(df_list, axis=0)
vendas = vendas.dropna(subset=['DAT_VENDA'])
vendas.sort_values(by=['DAT_VENDA'], inplace=True)
vendas

mask_2021 = vendas['DAT_VENDA']>='2021-01-01'
mask_2022 = vendas['DAT_VENDA']<='2022-01-01'
vendas[mask_2021 & mask_2022]

mask_2021 = vendas['DAT_VENDA']>='2021-01-01'
mask_2022 = vendas['DAT_VENDA']<='2022-05-05'
vendas[mask_2021 & mask_2022]

vendas['DAT_VENDA'].min()

---

### Itens

Compreende dados históricos. Semelhante/equivalente ao dataframe de vendas?

chunksize = 1*10 ** 6
for chunk in pd.read_csv(prefix + 'STAGING/relevanc_item_20220425.zip', sep=';', chunksize=chunksize):
    itens = chunk
    break
del chunk

itens = itens.dropna(subset=['dat_venda','cod_cliente'])
itens.sort_values(by=['dat_venda'], inplace=True)
itens

mask_2021 = itens['dat_venda']>='2021-01-01'
mask_2022 = itens['dat_venda']<='2022-01-01'
itens[mask_2021 & mask_2022]

mask_2021 = itens['dat_venda']>='2021-01-01'
mask_2022 = itens['dat_venda']<='2022-05-05'
itens[mask_2021 & mask_2022]

---

### Ativação Stix

Não compreende registros históricos.
Descontinuada.

chunksize = 1*10 ** 6
for chunk in pd.read_csv(prefix + 'STAGING/ativacao_resgate_stix_20220425.zip', sep=';', chunksize=chunksize):
    ativacao = chunk
    break
del chunk

ativacao = ativacao.dropna(subset=['cod_mes','cod_cliente'])
ativacao

ativacao['cod_mes'].value_counts()

---

### Descontos Stix

* DAT_ATIVA representa o dia em que o cliente fez uso do desconto ou recebeu uma notificação sobre o desconto?
* O dado será descontinuado.
* Em qual dataset seria possível mapear informações de uso de cupons/descontos?

descontos = pd.read_csv(prefix + dict_tables['desconto'][0], sep=';')
descontos = descontos.dropna(subset=['DAT_ATIVA'])
descontos.sort_values(by=['DAT_ATIVA'], inplace=True)
descontos

mask_2021 = descontos['DAT_ATIVA']>='2021-01-01'
mask_2022 = descontos['DAT_ATIVA']<='2022-01-01'
descontos[mask_2021 & mask_2022]

---

### Email

Indica ids validos para ativação.

df_list = []
for file in dict_tables['email']:
    try:
        local_df = pd.read_csv(prefix + file, sep=';', low_memory=False)
    except:
        local_df = pd.read_csv(prefix + file, sep='\n', low_memory=False)
        column_list = local_df.columns[0].split(';')
        local_df = local_df[local_df.columns[0]].str.split(';', expand=True)
        local_df.columns = column_list
    df_list.append(local_df)
    print('added file: ', file)
print('-'*10)

email = pd.concat(df_list, axis=0)
email

email['email_deliverability_status'].value_counts()

email['email_permission_status'].value_counts()

mask_2021 = email['dat_referencia']>='2021-01-01'
mask_2022 = email['dat_referencia']<='2022-01-01'
email[mask_2021 & mask_2022]

---

### Loja

loja = pd.read_csv(prefix + dict_tables['loja'][0], sep=';')
loja

---

### Produto

produto = pd.read_csv(prefix + dict_tables['produto'][0], sep=';')
produto
