==> gpa-poc-churn/notebooks/05-ia-predict.ipynb.ipynb <==
import os
import sys
import pytz
import glob
import joblib
import pandas as pd
from datetime import datetime as dt
from sklearn.pipeline import Pipeline

sys.path.append(os.path.dirname(os.getcwd()) + '/src')
import utils
import get_data
import pipeline_modules
from guara.modeling.supervised_modelz import *

def load_artifacts():
    
    try:
        parent_dir = os.path.dirname(os.getcwd())
        print(f'parent dir: {parent_dir}/artifacts/staging/')
        production_id = glob.glob(parent_dir+'/artifacts/staging/*')[0].split('/')[-1]
        print(f'Production Experiment: {production_id}')
    except: 
        print('No experiment in production, exiting...')
        exit()

    model_path = parent_dir + f'/artifacts/staging/{production_id}/model.joblib'
    fe_pipe_path = parent_dir+ f'/artifacts/staging/{production_id}/fe_pipeline.joblib'
    fs_pipe_path = parent_dir+ f'/artifacts/staging/{production_id}/fs_pipeline.joblib'

    model = joblib.load(model_path)
    print(f'Model Loaded')

    fe_pipeline = joblib.load(fe_pipe_path)
    print(f'Feature engineering pipeline loaded')

    fs_pipeline = joblib.load(fs_pipe_path)
    print(f'Feature selection pipeline loaded')

    production_id = production_id
    
    return model, fe_pipeline, fs_pipeline

model, fe_pipeline, fs_pipeline = load_artifacts()

model

fe_pipeline

fs_pipeline

---

df_val = pd.read_parquet('gs://gpa-churn/data/processed/steps/after_stix_30.parquet')

df_val.head()

def make_predictions(
    X, 
    model,
    fe_pipeline,
    fs_pipeline
    ):
        
    md = SupervisedModelz('lgbm', 'regression')
    md.model = model
    utils.log('Successfully loaded model')
    
    initial_shape = X.shape
    X = fe_pipeline.transform(X)
    X = fs_pipeline.transform(X)
    utils.log('Successfully applied pipelines')
    utils.log(f'X transformed: {initial_shape} -> {X.shape}')

    pred = md.model.predict(X)

    return pred

features = list(df_val.columns)
features = [i for i in features if i!= 'target']

pred = make_predictions(
    df_val[features], 
    model,
    fe_pipeline,
    fs_pipeline
    )

pred

from datetime import datetime
import uuid
def make_predictions_table(pred, X):
    
    Xpred = pd.DataFrame()
    Xpred['cod_cliente'] = X['cod_cliente'].copy()
    Xpred['churn_prediction'] = pred
    Xpred['prediction_time'] = datetime.now().strftime(format='%Y-%m-%d %H:%M:%S')
    Xpred['variant'] = f'20220527'
    Xpred['batch_id'] = str(uuid.uuid4())
    Xpred['model_status'] = 'poc'

    return Xpred

make_predictions_table(pred, df_val)

# for i in range(28,32):
#     os.system(f'gsutil cp gs://gpa-churn/data/processed/steps/after_stix_{i}.parquet gs://gpa-churn/data/processed/test/after_stix_{i}.parquet')

import get_data

from google.cloud import storage
def get_prediction_data(
    bucket:str='gpa-churn',
    prefix:str='data/processed/test/after_stix_'
    ):
    
    storage_client = storage.Client()
    obj_list = storage_client.list_blobs(bucket)
    obj_list = [i.name for i in obj_list if prefix in i.name]
    
    df_list = []
    for obj in obj_list:
        local_df = pd.read_parquet('gs://gpa-churn/'+obj)
        df_list.append(local_df)
        print(f'added {prefix}{obj}')

    df = pd.concat(df_list, axis=0)
    df.drop_duplicates(inplace=True)
    df.reset_index(drop=True, inplace=True)
    
    target = 'target'
    features = list(df.columns)
    features = [i for i in features if i != target]

    X = df[features]
    print('Successfully read test data.')
    print(f'X:{X.shape}')
    
    return X

get_prediction_data()

---

df = pd.read_parquet('gs://gpa-churn/data/processed/output/predictions.parquet')
df.head(40)
