==> gpa-poc-churn/notebooks/06-ia-geographic-features.ipynb <==
import os
import gc
import sys
import sqlite3
import numpy as np
import pandas as pd
import seaborn as sn
import matplotlib.pyplot as plt
from google.cloud import storage
from sqlalchemy import create_engine
from sklearn.pipeline import Pipeline

sys.path.append(os.path.dirname(os.getcwd())+'/src')
import pipeline_modules

prefix = 'gs://gpa-churn/'

# listing objects in blob storage bucket
storage_client = storage.Client()
obj_list = storage_client.list_blobs('gpa-churn')
obj_list = [i.name for i in obj_list if 'STAGING/' in i.name]

tables_dict = {
    'register':[i for i in obj_list if 'cadastro' in i],
    'stix':[i for i in obj_list if 'stix_optin' in i],
    'email':[i for i in obj_list if 'optin_email' in i],
    'items':[i for i in obj_list if 'relevanc_item' in i],
    'stores':[i for i in obj_list if 'relevanc_store' in i],
    'products':[i for i in obj_list if 'relevanc_product' in i],
    'activations':[i for i in obj_list if 'mov_vend_crm' in i]
}

def get_register_data(
    tables_dict:dict
    ):
    
    register = read_files_from_staging(
        tables_dict['register']
        )

    register.sort_values(by=['idcliente','datultatual'], inplace=True, ascending=True)
    register.drop_duplicates(subset=['idcliente','datultatual'], keep='last')

    register_cols_to_drop = ['codorigcliente','codorigcliente']
    register = register.drop(columns=register_cols_to_drop)
    register = register.dropna(subset=['idcliente','cidadecli','ufcli'], how='any', axis=0)

    register_cols_to_rename = {
        'idcliente':'cod_cliente',
        'datcadorigem':'data_cadastro',
        'codsexo':'sexo',
        'cidadecli':'cidade',
        'ufcli':'uf',
        'datnasccli':'data_nascimento',
        'datultatual':'data_registro'
    }
    register.rename(columns=register_cols_to_rename, inplace=True)

    register_cols_to_lower = ['cidade', 'uf']
    for var in register_cols_to_lower:
        register[var] = register[var].str.lower()

    register = register.drop_duplicates(subset=['cod_cliente','sexo','cidade','uf','data_nascimento'])
    register.reset_index(drop=True, inplace=True)
    
    return register

register = get_register_data(tables_dict)
register

register.value_counts('uf')

def get_region(x):    
    region_dict = {
        'se':['sp','mg','rj','es'],  #Sudeste
        's':['pr','sc','rs'],        #Sul
        'ne':['ma','ce','rn','pb','pe','al','se','ba','pi'], #Nordeste
        'n':['rr','ap','am','pa','ac','ro','to'],   #Norte
        'co':['mt','df','go','ms']    #Centro Oeste
        }
    
    for key in region_dict.keys():
        try:
            if x in region_dict[key]:
                return key
        except:
            pass
              
    return 'null'

register

register['region_dict'] = register['uf'].apply(lambda x:get_region(x))

register

register.value_counts('region_dict')

---

# %pip install geopy
# %pip install pycep_correios

from geopy.geocoders import Nominatim
from pycep_correios import get_address_from_cep, WebService

geolocator = Nominatim(user_agent="get_lat_long", timeout=1)
def get_lat_long(municipio):

    location = geolocator.geocode(municipio)
    return location.latitude, location.longitude

for i in ['uberlandia', 'patrocinio', 'goias', 'sao paulo', 'porto alegre']:
    print(get_lat_long(i))
