==> gpa-poc-churn/notebooks/08-ia-pipeline-predictions.ipynb <==
# !pip install jupyter_contrib_nbextensions
# !jupyter contrib nbextension install - user
# from jedi import settings
# settings.case_insensitive_completion = True

# # Install ai platform and kfp
# USER_FLAG = "--user"
# !pip3 install {USER_FLAG} google-cloud-aiplatform==1.3.0 --upgrade
# !pip3 install {USER_FLAG} kfp --upgrade
# !pip install google_cloud_pipeline_components

# !pip install kfp --upgrade

# !gcloud services enable compute.googleapis.com         \
#                        containerregistry.googleapis.com  \
#                        aiplatform.googleapis.com  \
#                        cloudbuild.googleapis.com \
#                        cloudfunctions.googleapis.com

from typing import NamedTuple
from kfp.v2 import dsl
from kfp.v2.dsl import (Artifact,
                        Dataset,
                        Input,
                        Model,
                        Output,
                        Metrics,
                        ClassificationMetrics,
                        component, 
                        OutputPath, 
                        InputPath)

from kfp.v2 import compiler
from google.cloud import bigquery
from google.cloud import aiplatform
from google.cloud.aiplatform import pipeline_jobs
from google_cloud_pipeline_components import aiplatform as gcc_aip

USER_FLAG = "--user"
#!gcloud auth login if needed

# Get projet name
shell_output=!gcloud config get-value project 2> /dev/null
PROJECT_ID=shell_output[0]
PROJECT_ID

# Set bucket name
BUCKET_NAME="gs://gpa-churn/artifacts"

# Create bucket
PIPELINE_ROOT = f"{BUCKET_NAME}/pipeline-vertexai/"
PIPELINE_ROOT

REGION="southamerica-east1"
REGION

---

## Creating pipeline components

@component(
    base_image="gcr.io/gpa-poc-001/churn-base-image-src-xgb@sha256:61db16ec13bba7d8023fff61329c6c28a7eb119f8f837fce4c09258776c16727",
    output_component_file="get_prediction_data.yaml"
)

def get_prediction_data(
    Xpred_: Output[Dataset],
    cod_cliente_: Output[Dataset],
    df_date_: Output[Dataset],
    path:str='gs://gpa-churn/data/processed/test/'
    ):
    
    import os
    import gc
    import sys
    import numpy as np
    import pandas as pd
    from google.cloud import storage
    
    # extracting bucket and path info from path
    #-------------------------------------------------------
    bucket = path.split('/')[2]
    path_ref = '/'.join(i for i in path.split('/')[3:-1])
    
    # reading dataframes in path folder
    #-------------------------------------------------------
    storage_client = storage.Client()
    obj_list = storage_client.list_blobs(bucket)
    obj_list = [i.name for i in obj_list if path_ref in i.name]
    obj_list = obj_list[1:]
    df_list = []
    for obj in obj_list:
        local_df = pd.read_parquet(f'gs://{bucket}/{obj}')
        df_list.append(local_df)
        print(f'added {path}{obj}')
        
    # concatenating df_list and saving cod_client column in an independent df
    #-------------------------------------------------------
    df = pd.concat(df_list, axis=0)
    df.drop_duplicates(inplace=True)
    df.reset_index(drop=True, inplace=True)
    cod_cliente = df[['cod_cliente']].copy()
    df_date = df[['date']].copy()
    
    # selecting valid columns
    #-------------------------------------------------------
    df.drop(columns=['cod_cliente'], inplace=True)
    target = 'target'
    features = list(df.columns)
    features = [i for i in features if i != target]
    Xpred = df[features]
    print('Successfully read prediction data')
    print('shapes:')
    print(f'xtrain:{Xpred.shape}')
    
    # saving output datasets in pipeline
    #-------------------------------------------------------
    Xpred.to_parquet(Xpred_.path + '.parquet', index=False, compression='gzip')
    cod_cliente.to_parquet(cod_cliente_.path + '.parquet', index=False, compression='gzip')
    df_date.to_parquet(df_date_.path + '.parquet', index=False, compression='gzip')

@component(
    base_image="gcr.io/gpa-poc-001/churn-base-image-src-xgb@sha256:61db16ec13bba7d8023fff61329c6c28a7eb119f8f837fce4c09258776c16727",
    output_component_file="load_artifacts.yaml"
)

def load_artifacts(
    fe_pipeline_: Output[Model],
    fs_pipeline_: Output[Model],
    endpoint_information_: Output[Model],
    path:str='gs://gpa-churn/artifacts/training_pipeline/production/'
    ):
    
    import os
    import sys
    import json
    import pytz
    import joblib
    import pandas as pd
    from io import BytesIO
    from datetime import datetime
    from google.cloud import storage
    from sklearn.pipeline import Pipeline

    sys.path.append('/usr/app/')
    sys.path.append('/usr/app/src')
    import src.pipeline_modules as pipeline_modules
    
    # extracting bucket and path info from prefix
    #-------------------------------------------------------
    bucket_name = path.split('/')[2]
    path_ref = '/'.join(i for i in path.split('/')[3:-1])
    
    # creating storage access point
    #-------------------------------------------------------
    storage_client = storage.Client()
    bucket = storage_client.get_bucket(bucket_name)
    
    # reading pipeline artifacts
    #-------------------------------------------------------
    pipe_dict = {
        'fe_pipeline':None,
        'fs_pipeline':None
    }

    for pipe in pipe_dict.keys():
        art_file=f'{path_ref}/{pipe}.joblib'
        blob = bucket.blob(art_file)
        art_obj = BytesIO()
        blob.download_to_file(art_obj)
        pipe_dict[pipe]=joblib.load(art_obj)
    
    # saving feature engineering and selection artifacts within pipeline
    #-------------------------------------------------------
    obj_list = [fe_pipeline_, fs_pipeline_]
    key_list = list(pipe_dict.keys())
    for i in range(len(obj_list)):
        file_name = obj_list[i].path + '.joblib'
        with open(file_name, 'wb') as file:
            joblib.dump(pipe_dict[key_list[i]], file)
    
    # reading endpoint_information artifact
    #-------------------------------------------------------
    blob = bucket.blob(f'{path_ref}/endpoint_information.json')
    endpoint_information = json.loads(blob.download_as_string(client=None))
    
    # saving endpoint_information artifact within pipeline
    #-------------------------------------------------------
    with open(endpoint_information_.path+'.json', 'w') as file:
        json.dump(endpoint_information, file)

@component(
    base_image="gcr.io/gpa-poc-001/churn-base-image-src-xgb@sha256:61db16ec13bba7d8023fff61329c6c28a7eb119f8f837fce4c09258776c16727",
    output_component_file="make_predictions.yaml"
)

def make_predictions(
    Xpred_: Input[Dataset],
    cod_cliente_: Input[Dataset],
    df_date_: Input[Dataset],
    fe_pipeline_: Input[Model],
    fs_pipeline_: Input[Model],
    endpoint_information_: Input[Model],
    predictions_df_: Output[Dataset],
    bucket:str='gpa-churn',
    output_path:str='data/processed/output/'
    ):
    
    import os
    import sys
    import json
    import pytz
    import uuid
    import joblib
    import pandas as pd
    import xgboost as xgb
    from datetime import datetime
    from google.cloud import storage
    from google.cloud import aiplatform
    from sklearn.pipeline import Pipeline

    sys.path.append('/usr/app/')
    sys.path.append('/usr/app/src')
    import src.utils as utils
    import src.pipeline_modules as pipeline_modules
    
    # loading artifacts
    #-------------------------------------------------------
    fe_pipeline = joblib.load(fe_pipeline_.path+'.joblib')
    fs_pipeline = joblib.load(fs_pipeline_.path+'.joblib')
    with open(endpoint_information_.path+'.json', 'r') as file:
        endpoint_information = json.load(file)
    project_id = endpoint_information['project_number']
    endpoint_id = endpoint_information['endpoint']
    
    # reading input arguments
    #-------------------------------------------------------
    Xpred = pd.read_parquet(Xpred_.path+'.parquet')
    cod_cliente = pd.read_parquet(cod_cliente_.path+'.parquet')
    df_date = pd.read_parquet(df_date_.path+'.parquet')
    
    # applying pipelines
    #-------------------------------------------------------
    Xpred = fe_pipeline.transform(Xpred)
    Xpred = fs_pipeline.transform(Xpred)
    
    # applying cod_cliente as index
    #-------------------------------------------------------
    Xpred = pd.concat([df_date, cod_cliente, Xpred], axis=1)
    Xpred.set_index(['date','cod_cliente'], inplace=True)
    
    # making predictions dictionary
    #-------------------------------------------------------
    predictions_dict = {
        'cod_cliente':[],
        'churn_prediction':[],
        'reference_date':[],
        'model_id':[]
    }
    prediction_time = datetime.now().strftime(format='%Y-%m-%d %H:%M:%S')
    endpoint = aiplatform.Endpoint(f'projects/{project_id}/locations/us-central1/endpoints/{endpoint_id}')
    for index, row in Xpred.iterrows():
        sample = list(row.values)
        pred = endpoint.predict([sample])
        predictions_dict['cod_cliente'].append(index[1])
        predictions_dict['churn_prediction'].append(pred[0][0])
        predictions_dict['reference_date'].append(index[0])
        predictions_dict['model_id'].append(pred[1])
    
    predictions_df = pd.DataFrame(predictions_dict)
    predictions_df['prediction_time'] = datetime.now().strftime(format='%Y-%m-%d %H:%M:%S')
    batch_id = str(uuid.uuid4())
    predictions_df['batch_id'] = batch_id
    predictions_df['model_stage'] = 'poc'
    
    # saving predictions dataframe in pipeline and in output_path
    #-------------------------------------------------------
    predictions_df.to_parquet(predictions_df_.path+'.parquet', index=False, compression='gzip')
    
    # upload predictions output to cloud storage
    #-------------------------------------------------------
    predictions_df.to_parquet('predictions.parquet', index=False, compression='gzip')
    storage_client = storage.Client()
    bucket = storage_client.get_bucket(bucket)
    storage_file= f'{output_path}{batch_id}.parquet'
    blob = bucket.blob(storage_file)
    blob.upload_from_filename('predictions.parquet')

---

# creating the pipeline
from datetime import datetime
timestamp=datetime.now().strftime("%Y%m%d%H%M%S")
pipeline_label = f'pipeline-churn-prediction-'

@dsl.pipeline(
    pipeline_root=PIPELINE_ROOT,
    name=pipeline_label)

def pipeline(
    bucket:str='gpa-churn',
    input_data_path:str='data/processed/drift_validation/',
    artifacts_path:str='artifacts/training_pipeline/production/',
    predictions_path:str='data/processed/output/'
    ):
    
    data_op = get_prediction_data(
        path=f'gs://{bucket}/{input_data_path}'
        )
    
    load_artifacts_op = load_artifacts(
        path=f'gs://{bucket}/{artifacts_path}'
        )
    
    make_predictions(
        data_op.outputs['Xpred_'],
        data_op.outputs['cod_cliente_'],
        data_op.outputs['df_date_'],
        load_artifacts_op.outputs['fe_pipeline_'],
        load_artifacts_op.outputs['fs_pipeline_'],
        load_artifacts_op.outputs['endpoint_information_'],
        bucket=bucket,
        output_path=predictions_path
        )

compiler.Compiler().compile(pipeline_func=pipeline,
        package_path='ml_pipeline_prediction.json')

---

start_pipeline = pipeline_jobs.PipelineJob(
    display_name=pipeline_label,
    template_path="ml_pipeline_prediction.json",
    enable_caching=True,
    location=REGION,
)

start_pipeline.run()
