==> gpa-poc-churn/notebooks/10-ia-solutiondemo.ipynb <==
-------
# Solution DEMO

import os
import sys
import json
import joblib
import pandas as pd
from io import BytesIO
from google.cloud import storage

---
### Reading artifacts

storage_client = storage.Client()
bucket = storage_client.get_bucket('gpa-churn')
path_ref = 'artifacts/training_pipeline/production/'

endpoint_information = 'endpoint_information.json'
blob = bucket.blob(f'{path_ref}{endpoint_information}')
endpoint_information = json.loads(blob.download_as_string(client=None))
endpoint_information

model_monitor_information = 'model_monitor_information.json'
blob = bucket.blob(f'{path_ref}{model_monitor_information}')
model_monitor_information = json.loads(blob.download_as_string(client=None))
model_monitor_information['model_monitor_id']

metrics = 'metrics.parquet'
metrics = pd.read_parquet(f'gs://gpa-churn/{path_ref}{metrics}')
metrics

---

## Endpoint (Real-time)

### Create input and upload to cloud storage

path = 'gs://gpa-churn/data/processed/input/'
bucket = 'gpa-churn'
path_ref = 'data/processed/input/'

# reading dataframes in path folder
#-------------------------------------------------------
storage_client = storage.Client()
obj_list = storage_client.list_blobs(bucket)
obj_list = [i.name for i in obj_list if path_ref in i.name]
obj_list = obj_list[1:5]
df_list = []
for obj in obj_list:
    local_df = pd.read_parquet(f'gs://{bucket}/{obj}')
    df_list.append(local_df)
    print(f'added {path}{obj}')

# concatenating df_list and saving cod_client column in an independent df
#-------------------------------------------------------
df_endpoint_input = pd.concat(df_list, axis=0)
df_endpoint_input.drop_duplicates(inplace=True)
df_endpoint_input.reset_index(drop=True, inplace=True)

df_endpoint_input.head()

df_endpoint_input[df_endpoint_input['idade']>=28000]
df_endpoint_input.sample(n=2000).to_parquet('gs://gpa-churn/data/processed/drift_validation/dataset_0.parquet', index=False, compression='gzip')

At this point, the trigger should start the endpoint prediction pipeline.

### Reading predictions from endpoint (real-time approach)

filename = '51f0d22d-867d-4d3a-b1f5-243bbd20faab.parquet'
path = 'gs://gpa-churn/data/processed/output/'
full_path = f'{path}{filename}'
df_endpoint_output = pd.read_parquet(full_path)
df_endpoint_output.head()

# unique client ids
len(df_endpoint_output)

---

## Batch

### Create input and upload to cloud storage

path = 'gs://gpa-churn/data/processed/input/'
bucket = 'gpa-churn'
path_ref = 'data/processed/input/'

# reading dataframes in path folder
#-------------------------------------------------------
storage_client = storage.Client()
obj_list = storage_client.list_blobs(bucket)
obj_list = [i.name for i in obj_list if path_ref in i.name]
obj = obj_list[1]
batch_filepath = f'gs://{bucket}/{obj}'
df_batch_input = pd.read_parquet(batch_filepath)
print(f'added {batch_filepath}')
df_batch_input.head()

df_batch_input.to_parquet(batch_filepath, index=False, compression='gzip')

At this point, the trigger should start the batch prediction pipeline.

### Reading batch prediction output

filename = 'predictions.parquet'
path = 'gs://gpa-churn/data/processed/batch_output/'
full_path = f'{path}{filename}'
df_batch_output = pd.read_parquet(full_path)
df_batch_output.head()

# unique client ids
len(df_batch_output)
